+++ app/src/main/java/com/android/xrayfa/di/GlobalModule.kt
@@ -5,15 +5,15 @@ import com.android.xrayfa.TrafficDetector
 import com.android.xrayfa.TrafficDetectorImpl
 import com.android.xrayfa.dao.LinkDao
 import com.android.xrayfa.dao.LinkDatabase
-import hev.htproxy.di.qualifier.Application
-import hev.htproxy.di.qualifier.Background
-import hev.htproxy.di.qualifier.Main
-import hev.htproxy.utils.NetPreferences
+import xrayfa.tun2socks.qualifier.Application
+import xrayfa.tun2socks.qualifier.Background
+import xrayfa.tun2socks.qualifier.Main
+import xrayfa.tun2socks.utils.NetPreferences
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
-import hev.htproxy.TProxyService
-import hev.htproxy.Tun2SocksService
+import xrayfa.tun2socks.TProxyService
+import xrayfa.tun2socks.Tun2SocksService
 import java.util.concurrent.Executor
 import java.util.concurrent.Executors
 import javax.inject.Singleton
