+++ build.gradle.kts
@@ -4,4 +4,5 @@ plugins {
     alias(libs.plugins.kotlin.android) apply false
     alias(libs.plugins.kotlin.compose) apply false
     kotlin("kapt") version "2.2.10" apply false
+    alias(libs.plugins.android.library) apply false
 }
