+++ common/build.gradle.kts
@@ -1,36 +1,18 @@
 plugins {
-    id("com.android.library")
+    alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
     kotlin("kapt") version "2.2.10"
 }
 
 android {
-    namespace = "hev.htproxy"
+    namespace = "com.android.xrayfa.common"
     compileSdk = 36
 
     defaultConfig {
-        //applicationId = "hev.htproxy"
         minSdk = 28
-        targetSdk = 35
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-//        ndk {
-//            abiFilters += listOf("armeabi-v7a","arm64-v8a","x86","x86_64")
-//        }
-        externalNativeBuild {
-            ndkBuild {
-                arguments  += listOf("APP_CFLAGS+=-DPKGNAME=xrayfa/tun2socks -ffile-prefix-map=${rootDir}=."
-                ,"APP_LDFLAGS+=-Wl,--build-id=none")
-            }
-        }
-    }
-
-
-
-    externalNativeBuild {
-        ndkBuild {
-            path = file("src/main/jni/Android.mk")
-        }
+        consumerProguardFiles("consumer-rules.pro")
     }
 
     buildTypes {
@@ -53,14 +35,13 @@ android {
 
 dependencies {
 
-    implementation(project(":common"))
+    implementation("androidx.datastore:datastore-preferences:1.1.7")
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
-
     implementation(libs.dagger)
     kapt(libs.dagger.compiler)
     implementation (libs.dagger.android)
