+++ tun2socks/src/main/java/xrayfa/tun2socks/Tun2SocksService.kt
@@ -1,4 +1,4 @@
-package hev.htproxy
+package xrayfa.tun2socks
 
 interface Tun2SocksService {
