+++ tun2socks/src/main/java/xrayfa/tun2socks/qualifier/Application.java
@@ -1,6 +1,7 @@
-package hev.htproxy.di.qualifier;
+package xrayfa.tun2socks.qualifier;
 
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -8,8 +9,8 @@ import java.lang.annotation.Target;
 
 import javax.inject.Qualifier;
 
-
 @Qualifier
+@Documented
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Main {
+public @interface Application {
 }
