+++ tun2socks/src/main/java/xrayfa/tun2socks/qualifier/Background.java
@@ -1,7 +1,6 @@
-package hev.htproxy.di.qualifier;
+package xrayfa.tun2socks.qualifier;
 
 
-import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -10,7 +9,6 @@ import java.lang.annotation.Target;
 import javax.inject.Qualifier;
 
 @Qualifier
-@Documented
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Application {
+public @interface Background {
 }
