+++ tun2socks/src/main/java/xrayfa/tun2socks/utils/NetPreferences.java
@@ -1,4 +1,4 @@
-package hev.htproxy.utils;
+package xrayfa.tun2socks.utils;
